Extract BinderAnchor and ResultAnchor from LambdaContainer
